Added authentication

+++ src/tasks/tasks.service.ts
@@ -9,6 +9,7 @@ import { CreateTaskDto } from './dto/create-task.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Task } from './task.entity';
 import { TaskStatus } from './task-status.enum';
+import { User } from 'src/auth/user.entity';
 @Injectable()
 export class TasksService {
   constructor(
@@ -16,11 +17,11 @@ export class TasksService {
     private taskRepository: TaskRepository,
   ) {}
 
-  async getTasks(filterDto: GetTasksFilterDto): Promise<Task[]> {
+  async getTasks(filterDto: GetTasksFilterDto, user: User): Promise<Task[]> {
     const { status, search } = filterDto;
     const query = this.taskRepository.createQueryBuilder('task');
 
-    // query.where('task.userId = :userId', { userId: user.id });
+    query.where('task.userId = :userId', { userId: user.id });
 
     if (status) {
       query.andWhere('task.status = :status', { status });
@@ -39,28 +40,9 @@ export class TasksService {
     }
   }
 
-  // getTasksWithFilter(filterDto: GetTasksFilterDto): Task[] {
-  //   const { status, search } = filterDto;
-
-  //   let tasks = this.getAllTasks();
-
-  //   if (status) {
-  //     tasks = tasks.filter((task) => task.status === status);
-  //   }
-
-  //   if (search) {
-  //     tasks = tasks.filter(
-  //       (task) =>
-  //         task.title.includes(search) || task.description.includes(search),
-  //     );
-  //   }
-
-  //   return tasks;
-  // }
-
-  async getTaskById(id: number): Promise<Task> {
+  async getTaskById(id: number, user: User): Promise<Task> {
     const task = await this.taskRepository.findOne({
-      where: { id },
+      where: { id, userId: user.id },
     });
     if (!task) {
       throw new NotFoundException(`Task: ${id} not found`);
@@ -69,27 +51,33 @@ export class TasksService {
     return task;
   }
 
-  async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-    const task = await this.getTaskById(id);
+  async updateTaskStatus(
+    id: number,
+    status: TaskStatus,
+    user: User,
+  ): Promise<Task> {
+    const task = await this.getTaskById(id, user);
     task.status = status;
     await task.save();
     return task;
   }
 
-  async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
+  async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
     const { title, description } = createTaskDto;
 
     const task = new Task();
+    task.user = user;
     task.title = title;
     task.description = description;
     task.status = TaskStatus.OPEN;
     await task.save();
+    delete task.user;
 
     return task;
   }
 
-  async deleteTask(id: number): Promise<void> {
-    const response = await this.taskRepository.delete(id);
+  async deleteTask(id: number, user: User): Promise<void> {
+    const response = await this.taskRepository.delete({ id, userId: user.id });
     if (response.affected === 0) {
       throw new NotFoundException(`Task: ${id} not found`);
     }
